requirements.txt
python=3.9.15
django=3.2
djangorestframework=3.12
mysql
mysqlclient=2.1
pillow=9.4 
django-cleanup==6.0